requirements.txt
PySmbClient

